+++ packages/tkati-node-el/src/tkati_node_el/settings.py
@@ -1,0 +1,8 @@
+from tkati_core.node import InputSettings, OutputSettings
+from tkati_core.settings import TomlBaseSettings
+
+
+class AppSettings(TomlBaseSettings):
+    input: InputSettings
+    output: OutputSettings
+    dlq: OutputSettings | None = None